Modify XBee.ts to provide send/receive data
This change adds the missing functionality to the XBee file that it
should send and receive MessagePack-encoded JSON. It now has a
sendData function that will take an object and destination address as
well as a callback onReceiveData function when it receives an object.

Again, all these objects must be MessagePack-encoded JSON strings.

Also, the comment enforcing rule in `.eslintrc.yml` has been removed.
It should be fine to make multiple line comments with the // notation.