Adds the UpdateHandler class
This change adds the existing UpdateHandler class from GCS to be used
in the communications API. This class is remade to be easier to read
for newer developers.

This class is in charge of creating instances of "handlers" whose job
is to look for certain events and perform two callback functions:
one is to simply run and the other is to determine whether the handler
should continue existing after that event or not.

Testing with mocha and chai is implemented with this class.