Refactor extra types to misc.ts
 ## Why is the change being made?

This change is made because it is easier to the developer to have all
these extra types in its own file. Also this change addresses the issue
that therea re no typeguards for some messages.

 ## What has changed to address the problem?

This change moves miscellaneous types to its own file `misc.ts`. This
change also adds typings to all the messages. There is no need to add
typings for tasks yet, waste of time to add (perhaps in the future).

 ## How was this change tested?

This change was tested with `npm test`.