Refactor UpdateHandler and add unit tests to it
 ## Why is the change being made?

This change is made because the object definition for
`UpdateHandler#handlers` is hard for developers to interpret.

 ## What has changed to address the problem?

This change changes that to a `Map<string, Handler[]` to make it easier
for developers to read. More unit testing has been added to
`UpdateHandler` to improve coverage.

 ## How was this change tested?

This change was tested with `npm test`.