Fix typeguards and modify JSONMessage typeguard
 ## Why is the change being made?

This change is made because two typeguards do not return the proper
typeguard syntax. This change also addresses the fact the JSONMessage
should check more than basic message requirements.

 ## What has changed to address the problem?

This change fixes the typos in the typeguards for StopMessage and
ConnectionAcknowledgementMessage. This change also adds additional
typechecks in JSONMessage's typeguard.

 ## How was this change tested?

This change was tested with `npm test`

 ## Related documents, URLs, commits