Fix package export of config
 ## Why is the change being made?

This change is made because when importing this package, getting config
variables would require a `config.default` call (ex. to get
`messageSendRateMs`, the variable is called
`config.default.messageSendRateMs` when it should be
`config.messageSendRateMs`.

 ## What has changed to address the problem?

This change changes the way `config.ts` exports its variables and how
`index.ts` imports its variables.

 ## How was this change tested?

This change was tested with `npm run test`.

 ## Related documents, URLs, commits