Update README instructions and add linting
 ## Why is the change being made?

This change is made because README has to state instructions on
including this package to Node.js projects. This package will not be put
on the npm website so all projects using this package must locally
depend on it.

This change decided that it is better for the developer to include this
package in the same folder as the project and simply link it to the
project. This way, updating code is as simple as pulling from the
branch, instead of having to perform a manual reinstallation of the
package.

 ## What has changed to address the problem?

This change adds instructions on how Node.js projects can use this
communications library locally by linking their projects to it.

This change also adds a linter for README.

 ## How was this change tested?

This change was tested with `npm test`.

 ## Related documents, URLs, commits