Add dependencies and typings for mock testing
 ## Why is the change being made?

This change is made because it is important to perform unit testing on
the XBee and Messenger classes to make sure they function properly. In
order to do this, the MockBinding class from `serialport` must be used.

 ## What has changed to address the problem?

This change performs multiple changes in order to add the MockBinding
class to the project:
1. Adds dependencies to `@serialport/stream`
`@serialport/binding-abstract` and `@serial/binding-mock`, and create
typings for each package (the website had the typings already but had no
file associated with it).
2. Moved new/existing package declaration files (.d.ts) to the following
filepath: `typings/package-name/index.d.ts`. Also added `paths` field to
`tsconfig.json` to point each package to its respective declaration
file.
3. Removed `test` in list of folders in `exclude` field in
`tsconfig.json`. This is to let files in `test` to access declaration
files under `typings`. The `main` and `types` fields are updated in
`package.json`.

Minor changes:
1. Removed `dist` and added instructions to build package before linking
it to README.
2. Minor refactoring in `XBee.ts` to make code more legible.

 ## How was this change tested?

This change was tested with `npm test`.

 ## Related documents, URLs, commits