Add webhooks for master builds to Discord
 ## Why is the change being made?

This change is made because it is important to put build status of the
`master` branch to Discord.

 ## What has changed to address the problem?

This change runs the webhooks script in `NGCP/GCS`. The script only
posts to the Discord channel if a build happens in `master` branch. The
`WEBHOOK_URL` has been set to the Discord webhook in Travis CI.

 ## How was this change tested?

This change was tested in a separate PR https://github.com/NGCP/GCS/pull/104

 ## Related documents, URLs, commits