Fix XBee and add unit testing to it
 ## Why is the change being made?

This change is made because the validity of XBee.ts is uncertain without
testing.

 ## What has changed to address the problem?

This change adds testing to XBee.ts through the XBee.test.ts file. It is
able to cover everything in the XBee class except the onReceiveObject
callback. This can be tested in the future.

This change fixes any issues in the XBee.ts file as a result of finding
bugs through testing, as well as making it more legible.

 ## How was this change tested?

This change was tested with `npm test`.

 ## Related documents, URLs, commits