Add sinon to mock unit testing
 ## Why is the change being made?

This change is made because coverage from previous commit is not that
great, and it is hard to test everything without mocks.

 ## What has changed to address the problem?

This change adds sinon, a mocking dependency to mock things for unit
testing. The XBee testing script now has one new script that mocks the
case in which the xbee fails to send data through sinon.

 ## How was this change tested?

This change was tested with `npm test`.

 ## Related documents, URLs, commits