Add unit tests for receiving data event
 ## Why is the change being made?

This change is made because this is the only situation that is not
tested for the XBee.

 ## What has changed to address the problem?

This change adds testing for the situation that the XBee receives any
type of frame. This goes deeply in al the cases that could occur from
what the XBee receives. To do this, the callback function from the event
was separated into its own function, made package-private and tested
there.

 ## How was this change tested?

This change was tested with `npm test`.

 ## Related documents, URLs, commits