Add UpdateHanlder#clearHandlers and simplify logic
 ## Why is the change being made?

This change is made because UpdateHandler requires a clearHandlers
function to remove all handlers (used from a required function
Messenger#stopSendingMessages).

 ## What has changed to address the problem?

This change adds the clearHandlers function to UpdateHandlers and adds
unit testing for it. This change also simplifies a lot of logic
concerning deleting handlers explicitly, and just calls the
handler.remove function instead.

This change also adds a throw statement if a handler is deleted but the
event does not exist.

 ## How was this change tested?

This change was tested with `npm test`.

 ## Related documents, URLs, commits