Refactor files for publishing
Prior to this change, this package was going to be kept local because it was exclusive to the NGCP project. However using `npm link` was not the ideal solution as that should only be used for development.

This change will modify certain files to prepare this package to be published to NPM. The package was renamed to `ngcp-communication-js`.